Attrempt to resolve the issue with encoding in DOS for the title of TOC:
The file "utf8-cyrillic.md" is utf8 encoded Cyrillic.

In cygwin the following command works properly -- the file is updated with TOC and its title is written correctly:

perl "../git-md-toc" -ut "[Сказки Пушкина]" -Tutf8 "utf8-cyrillic.md"

If the command is invoked from a script, the script itself should be in the proper encoding (for example, 866 or 1251). And the tool requires the --title-transcode option specifying the encoding of the file running the tool.

In DOS (WIN10) the command doesn't write the TOC title correctly. It always requires additional option for setting encoding for the title and this encoding SHOULDN'T be the same as the file encoding or a code page of the current DOS session. The code page defaults to 866 (however it could be any -- 866, 1251 or 65001) but --title-transcode MUST BE cp1251 always:

perl "..\git-md-toc" --title-transcode=cp1251 -ut "[Сказки Пушкина]" -Tutf8 "utf8-cyrillic.md"

I didn't research further. DOS is definitely painful puzzle!